requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
boto3==1.34.131
botocore==1.34.131
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
config==0.5.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
Django==4.2.13
django-cors-headers==4.4.0
djangorestframework==3.15.2
dnspython==2.2.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
faiss-cpu==1.8.0.post1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
greenlet==3.0.3
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain==0.2.5
langchain-aws==0.1.7
langchain-community==0.2.5
langchain-core==0.2.9
langchain-text-splitters==0.2.1
langsmith==0.1.81
lightgbm==4.4.0
MarkupSafe==2.1.1
marshmallow==3.21.3
matplotlib-inline==0.1.7
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.3
orjson==3.10.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.46
psutil==5.9.8
pure-eval==0.2.2
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.4.0
pymongo==4.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2022.2.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
smart-open==7.0.4
SQLAlchemy==2.0.31
sqlparse==0.5.0
tenacity==8.4.1
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.13
whitenoise==6.2.0
wrapt==1.16.0
xgboost==2.1.0
yarl==1.9.4
zipp==3.19.2
